refactor(gui): split project store support
